feature modules are created
