--- app/src/main/java/com/example/imagefinder/feature/imagedetail/ui/model/ImageDetailUiState.kt
@@ -1,11 +1,0 @@
-package com.example.imagefinder.feature.imagedetail.ui.model
-
-data class ImageDetailUiState(
-    val id: Long,
-    val tags: List<String>,
-    val userName: String,
-    val image: String,
-    val downloadsCount: Int,
-    val likesCount: Int,
-    val isBookmarked: Boolean
-)
