--- app/src/main/java/com/example/imagefinder/feature/route/ImageListDestination.kt
@@ -1,11 +1,0 @@
-package com.example.imagefinder.feature.route
-
-import com.example.imagefinder.common.navigation.BaseDestination
-import com.example.imagefinder.common.navigation.toRoute
-
-private const val IMAGE_LIST_ROUTE = "imageList"
-
-object ImageListDestination : BaseDestination() {
-    override val route: String
-        get() = toRoute(IMAGE_LIST_ROUTE)
-}
