+++ app/src/main/java/com/example/imagefinder/main/MainActivity.kt
@@ -1,0 +1,32 @@
+package com.example.imagefinder.main
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.Surface
+
+import androidx.compose.ui.Modifier
+import com.example.common.ui.theme.ImageFinderTheme
+import dagger.hilt.android.AndroidEntryPoint
+
+@AndroidEntryPoint
+class MainActivity : ComponentActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+        setContent {
+            ImageFinderTheme {
+                Surface(
+                    modifier = Modifier.fillMaxSize(),
+                    color = MaterialTheme.colors.background
+                ) {
+                    MainNavigation()
+                }
+            }
+        }
+    }
+
+}
