+++ core/common/build.gradle
@@ -51,12 +51,4 @@ dependencies {
     implementation libs.coil.compose
     implementation libs.androidx.compose.material
     implementation libs.androidx.compose.material3
-
-    //TODO : replace with version catalog
-    implementation 'androidx.core:core-ktx:1.10.1'
-    implementation 'androidx.appcompat:appcompat:1.6.1'
-    implementation 'com.google.android.material:material:1.9.0'
-    testImplementation 'junit:junit:4.13.2'
-    androidTestImplementation 'androidx.test.ext:junit:1.1.5'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
 }
