+++ core/common/src/main/res/drawable/img_placeholder_view.png
