+++ core/common/src/main/res/values/strings.xml
@@ -1,0 +1,26 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+
+    <string name="general_retry_action_label">Retry</string>
+    <string name="general_generic_error">Something is wrong</string>
+    <string name="general_connection_error">There is a problem with your network</string>
+
+    <!--    Content Description-->
+    <string name="cd_connection_view">Views</string>
+    <string name="cd_connection_downloads">Downloads</string>
+    <string name="cd_connection_like">Likes</string>
+
+    <!--    content description-->
+    <string name="simple_back_caption">Navigate back</string>
+
+    <string name="image_list_title">Your Images</string>
+    <string name="image_detail_title">Image Detail</string>
+
+    <string name="image_detail_downloaded">%d Users has downloaded this image</string>
+    <string name="image_detail_likes">%d Users has Liked this image</string>
+    <string name="image_detail_tags_title">Related Tags</string>
+
+    <string name="general_no_data_message">There is no data at the moment</string>
+
+
+</resources>
