+++ core/data/build.gradle
@@ -41,11 +41,4 @@ dependencies {
     implementation libs.hilt.android
     implementation libs.androidx.hilt.navigation.compose
     kapt libs.hilt.compiler
-
-    implementation 'androidx.core:core-ktx:1.10.1'
-    implementation 'androidx.appcompat:appcompat:1.6.1'
-    implementation 'com.google.android.material:material:1.9.0'
-    testImplementation 'junit:junit:4.13.2'
-    androidTestImplementation 'androidx.test.ext:junit:1.1.5'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
 }
