+++ core/navigation/build.gradle
@@ -1,12 +1,10 @@
 plugins {
-    id 'kotlin-kapt'
     id 'com.android.library'
     id 'org.jetbrains.kotlin.android'
-    id 'com.google.dagger.hilt.android'
 }
 
 android {
-    namespace 'com.example.database'
+    namespace 'com.example.navigation'
     compileSdk 33
 
     defaultConfig {
@@ -33,15 +31,5 @@ android {
 }
 
 dependencies {
-
-    implementation project(':core:model')
-
-    implementation libs.hilt.android
     implementation libs.androidx.hilt.navigation.compose
-    kapt libs.hilt.compiler
-
-    implementation libs.androidx.room.runtime
-    implementation libs.androidx.room.ktx
-    annotationProcessor libs.androidx.room.compiler
-    kapt libs.androidx.room.compiler
 }
