+++ core/navigation/src/main/java/com/example/navigation/route/navigation/BaseDestination.kt
@@ -1,0 +1,14 @@
+package com.example.navigation.route.navigation
+
+fun toRoute(vararg s: String) = s.joinToString("/")
+fun toPathWithParameter(vararg s: String) = s.joinToString("/") { "{$it}" }
+fun toPathWithOptionalParameter(vararg s: String) = s.joinToString("?") { "$it={$it}" }
+
+abstract class BaseDestination {
+    abstract val route: String
+    open val params: String = ""
+    val routeParams: String
+        get() = "$route/$params"
+    val routeOptionalParams: String
+        get() = "$route?$params"
+}
