+++ feature/imagelist/src/main/java/com/example/imagelist/navigation/ImageListGraph.kt
@@ -1,0 +1,20 @@
+package com.example.imagelist.navigation
+
+import androidx.navigation.NavGraphBuilder
+import androidx.navigation.NavHostController
+import androidx.navigation.compose.composable
+import com.example.imagelist.ui.ImageListScreen
+import com.example.navigation.route.ImageDetailDestination
+import com.example.navigation.route.ImageListDestination
+
+fun NavGraphBuilder.imageListGraph(
+    navController: NavHostController
+) {
+    composable(ImageListDestination.route) {
+        ImageListScreen(
+            onItemClick = { id ->
+                navController.navigate("${ImageDetailDestination.route}/$id")
+            }
+        )
+    }
+}
