+++ settings.gradle
@@ -21,3 +21,4 @@ include ':core:data'
 include ':core:model'
 include ':core:network'
 include ':core:database'
+include ':core:navigation'
